agent Вынес стейт в отдельный файл
